chatbot: isolate module responsibilities (chat_format + roles render/parse split)
Pull the model-facing wire shapes (ChatMessage/ChatRole/MessageToolCall(Function),
ToolDefinition) into a dedicated chat_format module — pure data, no logic. The
builders stay with the domain types (ConversationMessage/LLMResponse) and
ToolKind, which now import the shapes from chat_format.

Split roles.rs (which was doing three jobs) so each file is one responsibility:
- roles.rs       — the Role trait + contract types (RenderInputs, ParsedResponse,
                   ThinkingPolicy, FormatFlags)
- roles/render.rs — prompt assembly (minijinja, option-b)
- roles/parse.rs  — response decoding (regex)
- roles/primary_role.rs — PrimaryRole identity, delegates to render/parse

Dependency direction: conversation/tools -> chat_format <- roles. No behavior
change.
